Outlook-to-Excel
This script searches your inbox for an e-mail and grabs the body of that e-mail. and basically lets you do what you want with it. It can be slow if you have thousands of e-mails because it does unfortunately search through all of them.

I tried to find something similar through google but didnt have any luck. Apparently other people have been searching for something like this as well so here is my shot at it.

I'm sorry for the spaghetti code, i had to remove a lot of stuff from the script as the original script is well over 1000 lines of code at this point. I tried to make it as easy to read as possible and as easy to understand and edit. It can probably be a lot more optimized,.

The reason the Clean-Memory function is there is because i kept running out of memory, this can also probably be fixed by better code. :) But i'm not that good.

I didnt find a way to automatically have it select the next row in excel so if and elseif is used here.

If you have any questions or want to help make the script better please let me know :) Thank you